requirements.txt
groq>=0.9.0
mcp>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
requests>=2.31.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
bcrypt>=4.1.0
